// packages/plugins/plugin-calls/src/app/types/Messages.ts
//
// Copyright 2024 DXOS.org
//

export type User = {
  id: string;
  name: string;
  transceiverSessionId?: string;
  raisedHand: boolean;
  speaking: boolean;
  joined: boolean;
  tracks: {
    audio?: string;
    audioEnabled?: boolean;
    video?: string;
    videoEnabled?: boolean;
    screenshare?: string;
    screenShareEnabled?: boolean;
  };
};

export type RoomState = {
  meetingId?: string;
  users: User[];
};

export type ServerMessage =
  | {
      type: 'roomState';
      state: RoomState;
    }
  | {
      type: 'error';
      error?: string;
    }
  | {
      type: 'directMessage';
      from: string;
      message: string;
    }
  | {
      type: 'muteMic';
    }
  | {
      type: 'partyserver-pong';
    };

export type ClientMessage =
  | {
      type: 'userUpdate';
      user: User;
    }
  | {
      type: 'directMessage';
      to: string;
      message: string;
    }
  | {
      type: 'muteUser';
      id: string;
    }
  | {
      type: 'userLeft';
    }
  | {
      type: 'partyserver-ping';
    };